adds delegate analysts
merchant rev, invoice count, item count, pending
